EQDSKReader

Documentation for EQDSKReader.

EQDSKReader.EQDSKReader

EQDSKReader.Data

EQDSKReader.EQDSKReader — Module

EQDSKReader

Functionality to read EQDSK file.

From "G_EQDSK.pdf":

"...Briefly, a right-handed cylindrical coordinate system (R, φ, Ζ) is used. The G EQDSK provides information on the pressure, poloidal current function, q profile on a uniform flux grid from the magnetic axis to the plasma boundary and the poloidal flux function on the rectangular computation grid. Information on the plasma boundary and the surrounding limiter contour in also provided."

        character*10 case(6)
        dimension psirz(nw,nh),fpol(1),pres(1),ffprim(1),

        pprime(1),qpsi(1),rbbbs(1),zbbbs(1),

        rlim(1),zlim(1)
        c
        read (neqdsk,2000) (case(i),i=1,6),idum,nw,nh
        read (neqdsk,2020) rdim,zdim,rcentr,rleft,zmid
        read (neqdsk,2020) rmaxis,zmaxis,simag,sibry,bcentr
        read (neqdsk,2020) current,simag,xdum,rmaxis,xdum
        read (neqdsk,2020) zmaxis,xdum,sibry,xdum,xdum
        read (neqdsk,2020) (fpol(i),i=1,nw)
        read (neqdsk,2020) (pres(i),i=1,nw)
        read (neqdsk,2020) (ffprim(i),i=1,nw)
        read (neqdsk,2020) (pprime(i),i=1,nw)
        read (neqdsk,2020) ((psirz(i,j),i=1,nw),j=1,nh)
        read (neqdsk,2020) (qpsi(i),i=1,nw)
        read (neqdsk,2022) nbbbs,limitr
        read (neqdsk,2020) (rbbbs(i),zbbbs(i),i=1,nbbbs)
        read (neqdsk,2020) (rlim(i),zlim(i),i=1,limitr)
        c
        2000 format (6a8,3i4)
        2020 format (5e16.9)
        2022 format (2i5)

EQDSKReader.Data — Type

Data

Content structure of EQDSK file.

Attrs:

case    identification character string
nw      number of horizontal grid points
nh      number of vertical grid points
rdim    horizontal dimension in metter of computational box
zdim    vertical dimension in metter of computational box
rcentr  R in meter of vacuum toroidal magnetic field BCENTR
rleft   Minimum R in meter of rectangular computational box
zmid    Z of center of computational box in meter
rmaxis  R of magnetic axis in meter
zmaxis  Z of magnetic axis in meter
simag   poloidal flux at magnetic axis in Weber /rad
sibry   poloidal flux at the plasma boundary in Weber /rad
bcentr  Vacuum toroidal magnetic field in Tesla at RCENTR
current Plasma current in Ampere
fpol    Poloidal current function in m-T, F = RB_T on flux grid
pres    Plasma pressure in nt / m^2 on uniform flux grid
ffprim  FF’(ψ) in (mT)^2 / (Weber /rad) on uniform flux grid
pprime  P’(ψ) in (nt/m^2 ) / (Weber /rad) on uniform flux grid
psirz   Poloidal flux in Weber / rad on the rectangular grid points
qpsi    q values on uniform flux grid from axis to boundary
nbbbs   Number of boundary points
limitr  Number of limiter points
rbbbs   R of boundary points in meter
zbbbs   Z of boundary points in meter
rlim    R of surrounding limiter contour in meter
zlim    Z of surrounding limiter contour in mete